correct example NHC ligand IMes/SIMes
The previous definition of the IMes ligand was wrong.  The new
commit highlights the carbene character of the SIMes ligand by
extending a double bond to the ion to complex.  This appears to
be a pattern justified for instance both by the visual depiction
as well as the .mol file in ACS' set "common chemistry" about a
second generation Grubbs-Hoveyda catalyst (CAS 246047-72-3)[1]
with the saturated NHC ligand consulted today.

[1] https://commonchemistry.cas.org/

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>