drop prepending placeholder * in SMILES strings
By the previous commit, the (potential) ligands' stereogenic
centers are assigned an (R), or (S).  If missing, the placeholder
asterisk to indicate the dummy atom was prepended to the SMILES
string processed.  This is dropped because (R,R)-tartaric acid
then was depicted as (R,S)-tartaric acid.

It was tested to append the asterisk instead; while this preserved
the (R,R) configuration, this automatism now suggested a (primary)
binding via a hydroxyl group instead of the carboxylic acid or the
carboxylate though the later -- based on the pKa values -- would
appear more plausible.  In case of the box ligand, the preview even
suggested a bonding via one of the methyl groups.  Based on these
observations, neither a simple pre- nor append of the placeholder
seem suitable to generate the preview.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>